requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-etc==1.4.0
django-import-export==3.2.0
django-ipware==5.0.0
django-redis==5.3.0
django-rest-authemail==2.1.7
django-rq==2.8.1
django-secretballot==2.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.6
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
rq==1.15.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
